Update custom.py
skip rewrite

+++ jukebox/custom.py
@@ -198,13 +198,17 @@ def _legacy_load(f, map_location, pickle_module, **pickle_load_args):
             offset = f.tell()
         storage = deserialized_objects[key]
         s = str(key) + '.bin'
-        with open(s, 'wb') as f2:
-            for i in range(0, storage.size(), 8192):
-                f2.write(struct.pack(fmap[storage.dtype] * min(storage.size() - i, 8192), *(storage[i:i+8192])))
-        obj = storage.__class__.from_file(s, size=storage.size())
-        del storage
-        del deserialized_objects[key]
-        deserialized_objects[key] = obj
+        if not os.path.isfile(s):
+            with open(s, 'wb') as f2:
+                for i in range(0, storage.size(), 8192):
+                    f2.write(struct.pack(fmap[storage.dtype] * min(storage.size() - i, 8192), *(storage[i:i+8192])))
+        try:
+            obj = storage.__class__.from_file(s, size=storage.size())
+            del storage
+            del deserialized_objects[key]
+            deserialized_objects[key] = obj
+        except Exception:
+            deserialized_objects[key] = storage
         gc.collect()
     torch._utils._validate_loaded_sparse_tensors()
